refactor: export MaidrLiveApi type and unify context level discrimination
- Move the MaidrLiveApi interface into the liveData service and re-export
  it from maidr/react, so TypeScript consumers can type window.maidrLive
  without importing the side-effectful script-tag entry.
- Extract Context.isFigureLevel as the single source of truth for the
  figure/trace level discrimination used at construction, instruction
  resolution, and depth-1 stack restoration.
